feat: maxim plugin header support added to http transport
